+++ torch/csrc/jit/codegen/cuda/lower_utils.cpp
@@ -6,6 +6,7 @@
 #include <torch/csrc/jit/codegen/cuda/iter_visitor.h>
 #include <torch/csrc/jit/codegen/cuda/lower2device.h>
 #include <torch/csrc/jit/codegen/cuda/lower_thread_predicate.h>
+#include <torch/csrc/jit/codegen/cuda/kernel_ir_builder.h>
 
 #include <algorithm>
